implement HttpException methods on interface
