+++ src/Exception/HttpExceptionInterface.php
@@ -5,8 +5,11 @@ declare(strict_types=1);
 namespace SmartAssert\ApiClient\Exception;
 
 use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
 
-interface ActionExceptionInterface extends \Throwable, NamedRequestExceptionInterface
+interface HttpExceptionInterface extends \Throwable
 {
     public function getRequest(): RequestInterface;
+
+    public function getResponse(): ResponseInterface;
 }
